requirements.txt
appnope==0.1.0
awsebcli==3.12.1
beautifulsoup4==4.6.0
blessed==1.14.2
<<<<<<< HEAD
botocore==1.8.8
cachetools==2.0.1
=======
boto3==1.4.8
botocore==1.8.2
>>>>>>> origin/usermodel
cement==2.8.2
certifi==2017.11.5
chardet==3.0.4
click==6.7
colorama==0.3.7
decorator==4.1.2
Django==1.11.6
django-extensions==1.9.6
django-storages==1.6.5
djangorestframework==3.7.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==0.12.2
future==0.16.0
google-api-python-client==1.6.4
google-auth==1.2.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.10.3
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
oauth2client==4.1.2
oauthlib==2.0.6
parso==0.1.0
pathspec==0.5.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydub==0.20.0
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
<<<<<<< HEAD
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.12
=======
requests==2.9.1
s3transfer==0.1.11
>>>>>>> origin/usermodel
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.44.0
Werkzeug==0.12.2
